fixed form collapsing into navbar issue
